Move animation styling to toolbox
