+++ tool_animation/layer_styles/cell_difference.qml
