+++ tool_animation/layer_styles/node_current.qml
