+++ tool_animation/layer_styles/node_difference.qml
