+++ tool_animation/layer_styles/velocity_thresholds.qml
